03 [BookMyService Code] - Development/backend/requirements.txt
asgiref==3.4.1
asttokens==2.2.1
autopep8==1.5.7
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.1.2
datasketch==1.5.8
debugpy==1.6.5
decorator==5.1.1
Django==3.2.7
django-cors-headers==3.8.0
django-environ==0.9.0
django-phonenumber-field==7.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.2
entrypoints==0.4
executing==1.2.0
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
idna==3.4
ipykernel==6.19.4
ipython==8.8.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==7.4.8
jupyter_core==5.1.2
matplotlib-inline==0.1.6
mysql-connector-python==8.0.11
mysqlclient==2.1.0
nest-asyncio==1.5.6
numpy==1.24.1
packaging==22.0
pandas==1.5.2
parso==0.8.3
phonenumberslite==8.13.1
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.6.2
prompt-toolkit==3.0.36
protobuf==4.21.1
psutil==5.9.4
psycopg2==2.9.5
pure-eval==0.2.2
pycodestyle==2.7.0
Pygments==2.14.0
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==305
pyzmq==24.0.1
requests==2.28.1
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
sqlparse==0.4.2
stack-data==0.6.2
threadpoolctl==3.1.0
toml==0.10.2
tornado==6.2
traitlets==5.8.0
tzdata==2022.1
urllib3==1.26.12
wcwidth==0.2.5
whitenoise==6.4.0
